Feat/fix error catch

+++ functions/auth/Authentication.ps1
@@ -19,6 +19,7 @@ function Request-PiHoleAuth {
 
     catch {
         Write-Error -Message $_.Exception.Message
+        break
     }
 }
 
@@ -31,7 +32,7 @@ https://ftl.pi-hole.net/development-v6/docs/#get-/auth
 The URL to the PiHole Server, for example "http://pihole.domain.com:8080", or "http://192.168.1.100"
 
 .PARAMETER Password
-The API Password you generated from your PiHole server
+The API Password you generated from your PiHole v6 server
 
 .EXAMPLE
 Get-PiHoleCurrentAuthSession -PiHoleServer "http://pihole.domain.com:8080" -Password "fjdsjfldsjfkldjslafjskdl"
@@ -92,7 +93,9 @@ Get-PiHoleCurrentAuthSession -PiHoleServer "http://pihole.domain.com:8080" -Pass
     }
 
     finally {
-        Remove-PiHoleCurrentAuthSession -PiHoleServer $PiHoleServer -Sid $Sid
+        if ($Sid) {
+            Remove-PiHoleCurrentAuthSession -PiHoleServer $PiHoleServer -Sid $Sid
+        }
     }
 }
 
@@ -152,6 +155,7 @@ Get-PiHoleCurrentAuthSession -PiHoleServer "http://pihole.domain.com:8080" -Pass
         }
 
         Invoke-RestMethod @Params
+
     }
 
     catch {
@@ -159,7 +163,9 @@ Get-PiHoleCurrentAuthSession -PiHoleServer "http://pihole.domain.com:8080" -Pass
     }
 
     finally {
-        Remove-PiHoleCurrentAuthSession -PiHoleServer $PiHoleServer -Sid $Sid
+        if ($Sid) {
+            Remove-PiHoleCurrentAuthSession -PiHoleServer $PiHoleServer -Sid $Sid
+        }
     }
 }
